Create fewer bounding boxes.
Fixing a "bug" where the FeatureFloodCount object would create too many
overlapping bounding boxes. While this is not technically an error since
the bounding boxes taken as a whole still properly bounded the enclosed
feature, it likely resulted in a significant performance penalty for
many common grain growth simulations. This is a completely new and much
simplier algorithm.

refs #21772