Add robustness for fixed meshes in NearestNodeTransfer
This addresses the trace-trace coupling failure in blue crab that
cropped up after the merging of #20451. I believe this failure
started occurring due to the bounding box changes in #20451. But
anyway in general there will be no guarantee that you can safely
index into the solution vector for any given dof when the vector
is distributed amongst processes. So we must guard against that