Restoring the ability to swap between underlying containers
This PR just fixes a capability that was added a few years ago and has been a little neglected,
(e.g. making the data structures in the FeatureFloodCount object agnostic to the container type).
This could be used to do various performance analysis to continue to work on optimizations in the
future. It turns out there was only a single regression that was easy to resolve.

More importantly, I'm going to go ahead and flip the storage container from set to vector as even
though my testing is showing very little difference, vector does seem to consistently be a tiny bit
faster. More testing is needed.

refs #21891