Ensure destination setting works on all processes
The problem addressed in this commit was revealed in a phase
field split test, when in a two process run, process 1 did not have
any sub-apps. We should be able to construct `MooseAppCoordTransforms`
representing the non-translation transformations for both from and
to applications on all processes by broadcasting from rank 0 which,
I believe, should always have a sub-app. The changes in this commit
should help enable additition of coordinate transformation capabilities
to other transfer classes which query transformation information
for all global sub-apps