Consolidate parallel communication code in mortar constraints.
Provide motivation for the use of Cartesian Lagrange multipliers in the
context of mortar contact mechanics. The main reason this approach is
considered is for seamlessly using the variable condensation preconditioner,
which condenses the Lagrange multiplier based on global tractions. In order to
retrieve global tractions from the existing variationally consistent approach with
LMs based on the local frame, one needs to account for mortar geometry. This renders the
use of VCP impossible without a specific adaption of the preconditioner. For this reason,
and since it's also used in the literature within this variationally consistent approach, we
propose adding this variation in the definition of Lagrange multipliers.