Avoid dereferencing a nullptr (#21416)
When checking the subdomains covered in the constraint warehouse, we
try to dereference a mortar LM, which may not exist (e.g. a penalty mortar constraint).

Also, we may be interested in applying residuals with mortar pulling LMs but without solving
any constraint (as in the case of interface heat transfer). So, for all these reasons, we shouldn't
always assume that a mortar constraint object is solving a constraint equation by adding an LM.