Only perform size checking in devel/dbg modes
Looking at the code again I'm pretty sure that we only create this
decomposition once, so this check is probably not too expensive.
However, the error we emit is not going to be that much more meaningful
to a user than the PETSc error you get later. Moreover, this is truly
a developer error if we trigger the assertion, so I think an assertion
is totally appropriate