Turn assertion into error
I realized that it's very possible that a mistake in user input
can trigger nonmatching sizes between the split sum and the system
matrix size. So I've turned the assertion into a user-facing error.
I'm going to profile the error checking to see whether it's expensive
but I don't think it will be