Created const mesh getter function in Component
Added const and non-const mesh getter functions in Component.
Note that the const version could not overload the other due
to the compiler choosing the non-const version in non-const
member functions.

The non-const version checks that mesh setup has not been completed
and throws an error if it has. There should not be any mesh modification
past that point.

Component::_mesh still needs to be made private, but applications
need to be patched to use the new interfaces first.

Refs #22354