Made getter functions for THM problem in Component
Components will no longer be able to have a member function
reference to THMProblem because Components will be constructed
before the creation of THMProblem. Thus getter functions are
created to encapsulate this reference. Applications will then
be updated to use the getters instead of the reference. Then
the reference will be removed at that point.

Refs #22354