do not read output files to screen
When using PBS, do not print failed output to screen. While we probably
should print output to the screen for failed tests, the most used cases
are those from folks running on our HPC systems (a slow filesystem). By
printing the output filename/path instead of the output, vast amounts
of time is saved by not reading sometimes-hundreds-of-megabytes-of-output
from output files. This feature is negated when using --re, --verbose,
--failed-tests options.

Remove the deprecated queue options.