Add support for positions argument for higher order elemental variable
The positions was not added to the point request so the output was always taken at 0 0 0. This
showed wrong results for multiple subapps

The floating point operations (adding and subtracting the positions) also messed up the
caching map hashing and comparison, as the caching relies on points, eg 3 floating point numbers
- comparison relied on exact match seemingly, moved to absoluteFuzzyEqual
- hashing relied on exact match, now multiplies by 1e-8 and takes the int
The new hashing is a little rough. I m hoping others will have suggestions