Added error checking to mesh stitching
Libmesh's stitch_meshes and stitch_surfaces functions have an optional parameter
'enforce_all_nodes_match_on_boundaries' that, if true, throws an error if the
number of nodes on the specified boundaries don't match the number of nodes
that were merged.
This parameter is false by default.
This commit changes this parameter to true for all instances of stitch_meshes
and stitch_surfaces used in the framework.

Closes #23366