Automatic c_n/c_t for the mortar mechanical contact constraints (#17495).
Using the option 'automatic_c' allows to select c_n and c_t such that
c_n is roughly the inverse of a representative normal distance for contact
convergence. For example, 1/100 or 1/1000 the size of a finite element. And
c_t can be selected as the inverse of a representative sliding value over
a large step. The influence of JxW and the contact pressure is automatically
accounted for from previous step's quantities. Whereas contact pressure and
tangential relative velocity.