Skip initial residual evaluation if possible
This PR deprecates the parameter `compute_initial_residual_before_preset_bcs` in favor of
`use_pre_initial_residual`.

The user should set this parameter to true if they really want to use the pre-initial residual in the relative convergence check.

The pre-initial residual is defined as the norm of the residual evaluated at the beginning of the time step before any solution-modifying object is executed.
Solution-modifying objects are the objects that can potentially modify the solution vector. Currently, solution-modifying objects that are accounted for include:
- preset BCs
- constraints
- predictors

If the user does not explicitly ask to use the pre-initial residual, the pre-initial residual evaluation will be entirely skipped.
Even if the user asks for that, the pre-initial residual evaluation will be skipped if there is no active solution-modifying objects.
This behavior is consistent with the original discussion in the referenced issue.

See NonlinearSystemBase::shouldEvaluatePreinitialResidual() for the logic behind this. And please help me double check my logic.

close #10464