Skip initial residual evaluation if possible
This PR deprecates the parameter `compute_initial_residual_before_preset_bcs` in favor of
`use_pre_SMO_residual`.

The user should set this parameter to true if they really want to use the pre-initial residual in the relative convergence check.
The pre-SMO residual is defined as the norm of the residual evaluated at the beginning of the time step before any solution-modifying object is executed.

A legacy option is added in MooseApp to make the transition smoother. All modules except Peridynamics are using the new behavior.
Apps default to using the legacy behavior.

close #10464