Add stress base class to allow displaced mesh in UMAT routine.
The UMAT plugin is a special case of a stress calculator since strain and
displacement information is passed to this user routine. As such, we propose
modifying the hierarchy acknowledge this fact and handle mesh settings
independently.

UMAT can now be run on the displaced mesh if the user selects the parameter
use_displaced_mesh as 'true'.