Import StatusSystem in RunParallel
RunParallel uses the error status type, when the Tester object itself
does not enumerate that specific type.

Reasoning for not simply enumerating the 'error' type in Tester:
The Tester was never intended to exhibit 'error' types. That would be
too general; A test is supposed to pass or fail for specific reasons.

Closes #23721