Fix LowerDBlockFromSidesetGenerator functors
The TIMPI docs always said we require action functors to take *const*
lvalue references, but we seem to have never actually enforced that
until we added the rvalue reference support.

Refs https://github.com/libMesh/TIMPI/pull/113 , where Alex
describes making this change but I guess didn't actually push this
change?

This fixes https://github.com/libMesh/libmesh/pull/3500 for me, and
it'll be needed for the next libMesh submodule update too.

Refs #000