Fix test with no projection.
While migrating cases, I turn this lets-not-solve-edge correctly to
a lets-solve-edge correctly inadvertently. This problem does not converge
well when treating the contact patch edge properly. The influence of edge
dropping is known to mess with the condition number. In this case, while
the condition number with and without edge dropping treatment is similar,
convergence changes radically. So we will need to provide an alleviating
technique in addition to condensation, which we haven't used extensively
yet.