Limit explicitly how many objects can be created in mortar thermal action.
Users have no limit as to how many uos they can create to accumulate
its physics in the flux LM. But we are limiting the number of uos that
can be created internally by the action. This is not terrible as all known
(by me) production use cases do not have more than one conduction and one
radiation contributions (or others, but not repeated ones).