Clean up JSON output operations
We intended to support loading from JSON, but that never went anywhere.
If we want it in the future, we can specialize nlohmann::from_json.

Additionally, the to_json support in nlohmann achieves what we
wanted to do here with store/loadHelper, that is, these methods are
unneeded so also remove them

refs #24038