Patching the .la files so dependency paths make sense
This PR is working on MacOS, but only if I also set DYLD_LIBRARY_PATH to point to the location
where the libraries of the Multiapp is located. I've tracked this down a bit and the issue is
that Mac is ignoring the fact that I'm explicitly loading the dependent librarie (might be my fault)
and it's only paying attention to @rpath for where to look for dependent libraries even though I'm
explicitly trying to load them. However, we normally won't have the @rpaths of one application
stored in another (e.g. pronghorn doesn't know about our installed location of bison).

Bottom line, DYLD_LIBRARY_PATHS works for at least library interdependencies. I have not investigated this
on Linux yet