Address parallel issue with penalty mortar in contact action.
The relationship manager didn't get the chance to act if the user chose
mortar_penalty. This is fixed now.

I also removed the tangential velocity output because it gave different
output if distributed meshes were used for a 3D problem. This is strange
since the resulting displacements, tangential pressure, and accumulated
slip all passed. In fact the accumulated slip is computed from the tangential
velocity.