Back to install library archives
We are going to reinstall library archives, but this time we are also going to decide on
dynamic library uniqueness just based on archive file name (not the full path). This will mean
that we'll only get one version of "heat condition" for instance instead of two copies just because
they are in different directories.