Set up contact pairs in contact action using sideset proximity (#24122)
A new input parameter (nodeset_list) is added to the contact action.
For (pseudo) axisymmetric solids whose contact interactions are consistently
defined by wrap-around sidesets, the contact action can now loop over the
nodes in a sideset and obtain an approximate "center of gravity".

Distances between centers of gravity are computed. A minimum and reference
distance is computed to identify sidesets whose interaction is likely. As such,
those sidesets are arbitrarily given the primary and secondary labels. After than,
we remove the possibly repeated pairs.

This feature may be useful in cases for which many axisymmetric bodies for contact
need to be set up. Picture a large number of hollow cylinders that can come
into contact.