Template ModulesApp::registerAll
We need to template the modules registration method since when we dlopen open an
application we can wind up with multiple module loaders in memory each with that
method. Those methods however won't have the same information for different apps
due to all the conditional ifdefs for enabling different modules. By templating
we create unique entry points so we can find the right version corresponding to
our application. We'll keep the old method around and deprecate it.