Be more safe and prepare_for_use in mgs creating from scratch
Some follow-on mgs may not be checking whether the mesh is prepared
when they should be. So for safetey we can prepare more. Someday
when follow-on mgs are more consistently checking for mesh preparedness
when they need to be, then we can revert this commit (and also reduce
the number of prepares in other mgs creating meshes from scratch)