Update normal contact force for frictional node-face penalty (#24632)
This commit changes the way the normal contact force is used in the
node-face penalty friction algorithm. The test that failed with this
change only tests a point contact for a couple of time steps. That
failure was related to a too-large penalty factor. With the previous
approach, a too-large penalty factor may not initially have a negative
influence because the algorithm was using a cached normal force value.
Slightly decreasing that value, however, makes a more extended approach
to penalty contact work for the single point test modified in this commit.

Another example of a frictional cylinder is added to avoid regressions.
That test checks for tangential slips in opposite directions. The previous
formulation is unable to run this example.