Regold node-face penalty frictional tests.
- All tests appear to exercise mostly sliding
- Former penalty factor can be higher for these cases

The second item points probably to the fact that, before,
penalty was not implementing a truly penalty formulation and the
forces were being updated with previous iteration values, some sort
of simultaneous AL. If there is interest, an implementation like that
can be implemented with its own formulation enum, i.e. not 'PENALTY'.