Thermochimica variable checks in ChemicalCompositionAction (#24552)
Adds the checks for input parameters against Thermochimica system to `ChemicalCompositinAction`. After deleting the checks in `ThermochimicaNodalData`, this will allow us to create aux variables only for elements / phases / species in the Thermochimica system. The changes also allow us to provide an *ALL* option for the input parameters in which case the values obtained from Thermochimica will be used to create aux variables.

Need to write aux variables directly in UO, finish tokenizing.