Penalty implementation passes penalty tests
Small changes to code, tests, and gold files to make the penalty
implementation that enables augmented Lagrange work with existing
mortar penalty tests.

Small differences in the penalty friction implementation appear to
exist for stick behavior, but it is on the order of 1.0e-3. This
might be produced by the fact that for AL we cache the force instead
of the accumulated slip and some ensuing roundoff error.