Results are smooth with AL and initially verified against penalty.
Removed active set from the normal force. Active set information can be
retrieved from final contact pressure value e.g. (if > 0, active).

I noticed there is an artifact at the edge of the contact patch if the
penalty coefficient is very high. It looks as if there is a kink even though
the gap is identically satisfied. It may have to do with the mortar segment
mesh. We should keep an eye on this for other models.

I didn't follow exactly Uzawa's algorithm, so improvements are obviously possible.

We should:

 - Have more tests, bouncing block test and verification against LM formulation.
 - Take another look at Uzawa's algorithm. It is advertised as being superior in
   a system Newton loop, which is what we are doing.
 - Add convergence criterion for friction (the one now is fake/trivial) !!
 - Tests get stuck in parallel (revisit this).