Remove packaging requirement
Remove packaging requirement from mooseutils, and create our own.
Some of our build machines do not have the Python package: packaging
available.

Normally we would "SKIP" a test which requires a Python package not
being available. But in the case where we were using this package for
testing whether or not other python packages are available, we must
instead do it ourselves.