Ensure neighborhood node list is built on reinit for MOOSE::Iteration (#24801)
MOOSE::Iteration update strategy relies on building neighborhood lists at the
initialization stage. When mesh has initial adaptivity, doing this on initialization
is not enough as evidenced by the fact that neighborhood nodes are not found on the list
of ghosted nodes. If we do this on reinitialization, mesh information appeared to have
been updated with refined elements.

This is intended to enable initial adaptivity with Iteration patch update strategy in
parallel runs.