Address Guillaume's review.
The flag to build the set of neighborhood nodes is controlled from
the local reinit() method instead of just ensuring a second pass.

Also regold a test that yielded small differences after the update in this
commit. This test did not exactly match the 'always' patch update strategy results.
This latter approach is supposed to be more exact than iteration, so I take it
as the reference. After the change in this commit (i.e. building neighborhood lists
in every call to reinit instead of on construction) this regolded test that uses
'iteration' strategy yields the same results as the 'always' strategy. It looks as
if something got fixed by doing this.