Updated FindHungProcessTester.py to work on linux
Test currently does run on linux. Need to now add pattern recognition of output to verify that find_hung_process.py is working correctly.

As a note, sleep for 5 seconds to make sure that program is hung in a state that is expected. In the first few seconds bad_mpi.cpp does not have the expected output because of timers. There could be improvements later to make the program hang quicker, and stay in the hung state that we're looking for.

It may not be ideal to use pkill, but it does work for now since the name of the binary shouldn't change in the tester.