Allow per-iteration output in SteadyAndAdjoint solve from optimization solver (#25009)
This is a mildly invasive way of forcing proper Exodus iteration output in
steady and adjoint solves. We are casting in the Exodus object and passing
iteration number information back and forth from main app to child app.

There is a test that exemplifies the usage of this feature.

Restricts most changes to optimization module.