Ifndef directive for some Thermochimica errors (#24552)
Several mooseErrors in `ThermochimicaNodalData` are to safegaurd against possible deprecations / regressions in Thermochimica or `ChemicalCompositionAction`. They cannot be tested currently because variables are created automatically by `ChemicalCompositionAction` and possible errors are caught at that point. To hit the errors in ThermochimicaNodalData would require internal Thermochimica errors which we cannot "create" right now. To avoid failing code coverage due to this, enclosing these tests in `#ifndef NDEBUG`