For penalty, use "normalized" normal penalty value.
That is, in addition to get the physical gap, we divide by a value given
by normalize_c, which is a proxy for the physical area that surrounds that
node. This is done at the request of users that are familiarized with the
normalized_penalty option of node-face. The values would not be exactly the
same, but it should allow for using the same penalty value in both formulations.

For a mechanics problem, the normal penalty coefficient can be selected on the order
of the modulus of elasticity of the materials. In the case of AL, it should be
selected as the softest of the materials in contact.

Also:

Adjust tolerances and use absolute slip.

Tolerances are adjusted because there are nonzero quantities that
range from 1.0e-4 to 1.0e5, such as pressure, so we add a abs zero to
those tests.