Added bad_mpi source
This is the program that will hang a number <n> stacks deep.

The depth is specified with using the -s option. The "s" is short for stack. This program is used to test the accuracy of find_hung_process.py, and to also test the script when it is not on a job queuing system.

Later on this can be used to test the search algorithm in find_hung_process.py. The stack depth can be adjusted, and then the output from the script will tell you if it correctly narrowed in on the location of where the program actually hung (which stack).