Dependency constraint adjustments
This allows us to drop moose-tools -> moose-mpich dependency I
thought we were going to need.

It is `clang-tools` which seems to spur a lot of conflicts. So by
NOT constraining tools and test-tools to a specific clang-tools
version, we can allow any clang-tools version to be install which is
supported by moose-mpich (but not _installed_ by moose-mpich).

While we're discusing... it is advisable to never place a version
constraint on any package in our tools and test-tools package...
Except of course, Python and the tools -> test-tools dependency.