Extend per-iteration output to plain Steady solves.
This commit addresses Lynn's request to have this capability for
Steady solves as well. As of this commit, the Exodus object that
enables the per-iteration output is ExodusOptimizationSteady, which
would work for Steady and SteadyAndAdjoint solves.

The ability for the user to try to output per-iteration is restricted
to the optimization module.