Avoid ambiguous subdomain when stitching in test
libMesh stitch_meshes is going to start handling MOOSE "subdomain names
are first-class objects, ids might be just autogenerated" behavior
safely, in addition to the old libMesh "subdomain ids are first-class
objects, names are just expected to be matching" behavior.  But in
between those categories of case is "one submesh may be treating names
as first-class objects, another submesh has a matching id without any
name", and we can't figure out any safe thing to do there, so by default
now we'll scream and die when we see it.

So we definitely don't want to see it in CI.

This is all in service of catching bugs like in #25070, via
https://github.com/libMesh/libmesh/pull/3613