Fix conflicts on Linux
ICU and Autotools packages, appear to be the culprit somehow. Need to
pin these at specific versions. The bump to later MPICH requires
autotools bump. While doing so allowed a later version of ICU to
enter the mix, while all the peacock dependencies were still using
an older ICU package version. And because we have `strict` turned
on, moose-mpich would disallow the downgrade.