versioner.yaml to track build.sh
I am questioning why this was not already added. Its major, so we
must have had a reason. And it is clear that through the build.sh
script, major changes can occur in the resulting package.

I was about to leave it, but I see we're tracking both:

  update_and_rebuild_petsci.sh|libmesh.sh
and
  configure_petsc|libmesh.sh

So why not build.sh?