Allow execute on INITIAL, added test, fixed test
After discussion with other developers, we allow INITIAL as a valid execute_on
option. The default value is still set to TIMESTEP_END. All other value are
disallowed and will raise a ParamError.

A RunException test called 'execute_on_param_error' was added to test for the
above mentioned ParamError.

Finally, the 'signaling_nan_opt' test in fluid_properties was modified to
not be run in --recover mode because part 1 of the test crashes immediately
without writing a checkpoint (does not perform a timestep). Thus, the --recover
simulation crashes because it can't find a checkpoint file. Hence,
'recover = False' was added to the test.