Fix `nanoflann::searchParameters` call in `AutomaticMortarGeneration.C` (#25904)
The `searchParameters()` syntax in Nanoflann 1.5.0 does not have the obsolete `int checks_IGNORED_` param which was part of `searchParams()` arguments. This would cause `nanoflann::searchParameters(10)` calls in the MOOSE patch (https://github.com/idaholab/moose/pull/25962) to result in failing tests when using the nanoflann 1.5.0. This patch should allow those tests to pass.