Removed unnecessary PP outputs warning
The warning in question was redundant and less useful than
a check already being made in
AdvancedOutput::initPostprocessorOrVectorPostprocessorLists
but the check was broken in some cases because of _advanced_execute_on
not being consistent.

In AdvancedOutput::init, checks were being performed before
_advanced_execute_on was made consistent with the input parameters,
so the calls to initExecutionTypes was moved up before these
checks, since it has no dependence on the preceding calls in init().

Closes #25909