Install clang from system package manager
Install clang with dnf. Install matching gfortran version that
coincides with the version of gcc that gets installed as a
dependency from installing clang.

Set necessary /opt/rh/ toolchain paths. This was a key piece of
information I missed as the dependencies above were being
installed. `clang` gets installed to /usr/bin/, while the dependecies
in order to properly make use of said clang are installed in
/opt/rh/etc (some path that is not available by default like
/usr/bin is). Hence, confusion.

Closes #25831