Support old Nanoflann API *and* new Nanoflann API
This gets us around the chicken-and-egg problem of not being able to get
the Nanoflann update in libMesh to pass MOOSE tests.

Remember to build backwards compatibility into your library API updates,
kids!

We can remove the `namespace nanoflann` shims after the update makes it
into the libMesh submodule.

Refs #25904, but we can wait until the shims are gone before we close
that ticket.