Patched logic bug for update of _last_output_time.
Output objects have an attribute _last_output_time describing the simulation
time at which output last occured. Before this commit, the updating of this
variable occurred before checking the status of shouldOutput, resulting in
erroneous variable updates when shouldOutput is false. This commit moves the
variable update into the check of shouldOutput, correcting the issue.