Switch to conda-build, Add dep information
Switch from `mamba mambabuild` to `conda-build` in the generate script.
We will someday do the same for Civet recipes...

Add dependency information to empty_shell template. This will be
necessary for an upcoming change to ncrc-client. This increases the
package size by a minimal amount (all the meta data about what this
package requires, will now be in this package). The time it takes
to build this package also seems to be unaffected.

Refs #26343