Added wall time checkpoint, code cleanup
Added the capability that a checkpoint is written every 10 minutes of wall time
that a simulation runs (this wall time interval can be changed through the new
“wall_time_interval” input parameter).

Additionally, common code from the parent Output object that was duplicated in
the child Checkpoint object was refactored into Output.

Finally, some variable names were renamed for improved clarity.
The “interval” input parameter was deprecated in favor of “time_step_interval”
and the “minimum_time_interval” input parameter was deprecated in favor of
“min_simulation_time_interval”.
These parameter names have correspondingly been updated in all tests.