Added tests for wall time checkpoints
This commit adds test cases for the new wall time checkpoint functionality.
To test that wall time checkpoints work without the user creating a checkpoint
block or using the shortcut syntax, the command line parameter
‘output-wall-time-interval’ (and supporting code) was added to MooseApp as a
way to override the default wall time interval of 10 minutes without defining a
checkpoint in the input file. The 10 minute default value needed to be
overridden to a much shorter interval suitable for the test suit.