Include option to overwrite current solution.
This option is more consistent with the internal energy of the solid.

I.e., it overwrites the position of the contacting boundaries with the
result of the iterative contact algorithm. However, it sometimes results in
some sort of element inversion or impossible mappings. This needs to be
investigated/debugged. It may be due to MOOSE's order of operations.