Cleaned up logic in SignalTester.
The main part of this commit was reworking the logic in SignalTester.send_signal.
Namely, the check to see if the moose binary has started running and writing to the
output file was modified from an approach that copied and read the output file to
a much cheaper approach that simply queried the current position in the output file.

Additionally, the 'signal' parameter previously did not do anything. This has been patched.

Finally, the 'sleep_time' parameter has been removed because it was not deemed necessary to
the testing process (there should be no such thing as a signal being sent too early after the
binary has started to run; the checkpoint will simply be output at the end of the next time step).