Robustify Rayleigh Bernard two phase test
Refs failures from the combination of #26462 and #26227

To avoid negative viscosity interpolations I shrank the growth
factor from 2 to 1.5. Once I did this I got a CSVDiff which signaled
to me that the original result was solved to too loose a tolerance.
So I removed the steady state detection, which has potential issues
with it (see #26312), and extended the `end_time` until we hit
a fairly tight absolute tolerance