Go with odd number of elements
I noticed that when a diff occurred, looking at vel_y, it was a
mirror plane switch across the x = L/2 line. So by going with an
odd number of elements this doesn't occur. I tested this with
parallel and/or thread sweeps with and without distributed mesh
and could not get it to fail