Remove testing redundancy in JSON TestSearch
Each invocation of an oprof MOOSE executable with `--json` takes
about 5 seconds. It's far far worse when compiled with `-O0`. Both
of the queries in the `TestSearch` test were looking at "blocks" type
data. So there is no need to make this test run 2X as long to test
the same thing 2X