Patched checkpoint directory naming bug
Moose no longer allows multiple checkpoint objects. Because of this, input files
with user-specified checkpoints are no longer overwritten with the shortcut
`Outputs/checkpoint=true` when using the --test_checkpoint-half-transient flag
(previously known as the --half-transient flag). This changed enabled the
discovery of a bug when running moose in --recover mode: with checkpoint objects
named anything other than `out` moose was unable to find the checkpoint directory.
This is because moose was getting an empty value for the `file_base` parameter
in the checkpoint object before it was later set during the execution of the
Executioner. This commit patches this issue, ensuring the correct checkpoint
directory name is found.