Use boomeramg for hybridized diffusion tests
So the matrix is nonsymmetric which I need to investigate since the
corresponding libMesh example does have a symmetric matrix. My guess
is to do with the single face hybridizable, which is what the libMesh
example is doing, but the MOOSE example is not. However, the MOOSE
test still appears to be SPD "enough" for boomeramg to work well