Patched multiple procs writing to same file
Console objects have a parameter 'output_file', that when true, writes console
output to a file. Parallel simulations would write information from all
processors to the same file, which can lead to bad behavior. This commit
introduces a patch that only writes console content from the head proc to the
file. In the case that --keep-cout is passed to the executable, each proc writes
its console content to a separate file. The same bahavior can be achieved
without setting output_file to true by instead passing
'--keep-cout --redirect-stdout' to the executable.

Closes Ref #26174.