Update solid mechanics AD return mapping test
When computing scaling once we now regard a failure on the initial
time step during the scaling calculation as not satisfying a computed
scaling criteria. This means that we keep attempting an auto-scaling
calculation, after cutting the time-step, until we get an exceptionless
scaling calculation. This is different from the past where even if
something bad happened during the scaling calculation we accepted
the partial results of the scaling calculation.

Differences in old and new scaling behavior led to a different solve
history for this test. So now I am just manually supplying the scaling
factors to reproduce the old behavior