Much more straightforward try-catch design
The design is to put the `try` at the top of each method. The `catch` should
dispatch to `handleException`. An outer `try-catch` can also be applied and
should be applied when we are in an outer method like
`computeResidualTags/computeJacobianTags/computeResidualAndJacobian`. When we
are in an "inner" but generic method like `computeUserObjectsInternal` there
should not be an outer `try-catch` because we must allow the exception coming
from `handleException` to propagate up higher to a place where we know whether
we can appropriately handle it and cut the timestep else error out