Check for first local rank to avoid duplicate sum
This commit patches the bug reported in #27047, where the MultiAppVectorPostprocessorTransfer
gives incorrect results when pp values are transfered from subapps distributed
over more than 1 proc to the parent app. The transfer is executed by defining on
each proc a vector of zeros with length equal to the number of subapps. The
intention is that each subapp will put its pp value into the corresponding entry
of this vector, leaving the other entries as zero. In order to fill the remaining
zeros on each proc with the pp values from subapps not on the given proc, the
vector values are summed element-wise accross all processors. The old implementation
of this approach uses hasLocalApp(i) to check if subapp i belongs to the current
processor. If so, the corresponding vector value is changed from 0 to the pp value.
When a subapp is distributed accross more than one processor, this approach does
not work because hasLocalApp(i) returns true on multiple processors, thus the
vector that holds pp values gets the value for the distributed subapp on all
processors that it lives on. These values are added together, resulting in an
overall incorrect pp value transfered to the vpp. The solution to this issue is,
in addition to checking hasLocalApp(i) to see if subapp i lives on the current
processors, also check isFirstLocalRank() to determine if the current processor
is the first proc that subapp i is distributed on. This avoids the erroneous
duplication and summation of the same subapp's pp value.

Additionally, the test for the MultiAppVectorPostprocessorTransfer was updated
to distribute each subapp over at least 2 processors so issues like these are
not overlooked in the future.