Added bounds checking to getBoundaryID.
This commit addresses #17175, which reports an issue with the
RenameBoundaryGenerator. When the 'new_boundary' parameter value is given as
boundary id values instead of boundary names, and the new ids are greater than
2^15, type overflow occurs and the boundary is renamed to the largest possible
value for type BoundaryID instead of what the user requeted. This commit adds
bounds checking to user-requested boundary ids and emits a descriptive error
if the user requests to use bpundary id values outside the range of the
BoundaryID type.